requirements.txt
Pillow==2.6.1
RPi.GPIO==0.6.3
RTIMULib==7.2.1
argparse==1.2.1
arrow==0.4.2
bottle==0.12.9
chardet==2.3.0
colorama==0.3.2
gpiozero==1.0.0
html5lib==0.999
icalendar==3.9.1
ics==0.3.1
lxkeymap==0.1
mcpi==0.1.1
ndg-httpsclient==0.3.2
numpy==1.8.2
picamera==1.10
pifacecommon==4.2.1
pifacedigitalio==3.1.0
pyOpenSSL==0.13.1
pyasn1==0.1.7
pyfscache==0.9.12
pygame==1.9.2a0
pygobject==3.14.0
pyserial==2.6
python-dateutil==2.4.2
pytz==2015.7
requests==2.4.3
sense-hat==2.1.0
simple-cache==0.35
six==1.8.0
spidev==3.0
urllib3==1.9.1
virtualenv==15.1.0
wheel==0.24.0
wsgiref==0.1.2
